Split build info function proto and defs
This is to avoid duplicating the information contained in the
binaries. The way it's done now, the information is stored in the
binaries and won't be in the header file.

Only the function prototypes header file is installed if explicitly
requested. The header file containing the definitions is only for
building and will never be installed.